+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
+## [0.3.1] - 2021-09-02
+### Added
+- new Api: touchBus
+### Changed
+- it's not neccessary any more to specify the bus's name while calling createBus and getBus
+- support to pass a object parameter while calling ctx.loadJs and ctx.loadCss
 ## [0.3.0] - 2021-08-18
 ### Changed
 - remove the Bus's property `maxDependencyDepth` and `loadScriptByFetch`, use the new API `bus.config` instead